Add T* (userdata) as type. Static map:   lua_State* x T -> const char*
Returning from stack: if a T* is returned it is cast to
a void* and pushed as a lightuserdata (losses type information...).
Since userdata (with custom metatable) assumes Lua owns the memory
it's a problem to use try to associate a type with a lightuserdata.

Pushing to stack: if a userdata is on the stack it should be from
(owned by) the lua_State and so testing requires we get from
the lua_State instead of push the value (we would have to push
as a lightuserdata and would loss the type).

The optional stuff and returns aren't tested yet.

Maybe we can do more with a light lightuserdata wrapper...
though this may be too much.